# backend/scripts/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "actual_duration": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Duration"
          },
          "classification_method": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Method"
          },
          "created_via": {
            "anyOf": [
              {
                "pattern": "^(voice|text|manual|imported)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Via"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "effectiveness_rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectiveness Rating"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "energy_level": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Level"
          },
          "is_all_day": {
            "default": false,
            "title": "Is All Day",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_outcome": {
            "anyOf": [
              {
                "pattern": "^(productive|waste|neutral)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Outcome"
          },
          "planned_duration": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Duration"
          },
          "priority_confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Confidence"
          },
          "priority_level": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Level"
          },
          "recurrence_rule": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Rule"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "start_time",
          "end_time"
        ],
        "title": "EventCreate",
        "type": "object"
      },
...
      "EventResponse": {
        "properties": {
          "actual_duration": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Duration"
          },
          "classification_method": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "manual",
            "title": "Classification Method"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_via": {
            "anyOf": [
              {
                "pattern": "^(voice|text|manual|imported)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "manual",
            "title": "Created Via"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "effectiveness_rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 3,
            "title": "Effectiveness Rating"
          },
          "end_time": {
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "energy_level": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 3,
            "title": "Energy Level"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_all_day": {
            "default": false,
            "title": "Is All Day",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_outcome": {
            "anyOf": [
              {
                "pattern": "^(productive|waste|neutral)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "neutral",
            "title": "Meeting Outcome"
          },
          "planned_duration": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Duration"
          },
          "priority_confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.0,
            "title": "Priority Confidence"
          },
          "priority_level": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 3,
            "title": "Priority Level"
          },
          "recurrence_rule": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Rule"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "title",
          "start_time",
          "end_time",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "EventResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-Powered Smart Calendar System with BERT Priority Classification",
    "title": "KairoCal API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events/": {
      "get": {
        "description": "List events for the specified user with enhanced priority filtering and statistics\n\nEnhanced features:\n- Filter by priority ranges or specific levels\n- Filter by classification method (BERT vs manual vs rule-based)\n- Sort by priority, confidence, or time\n- Include priority distribution statistics\n- BERT-enhanced metadata in responses",
        "operationId": "list_my_events_api_v1_events__get",
        "parameters": [
          {
            "in": "query",
            "name": "cognito_sub",
            "required": true,
            "schema": {
              "title": "Cognito Sub",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "description": "Minimum priority level (1=Critical, 5=Very Low)",
            "in": "query",
            "name": "min_priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum priority level (1=Critical, 5=Very Low)",
              "title": "Min Priority"
            }
          },
          {
            "description": "Maximum priority level (1=Critical, 5=Very Low)",
            "in": "query",
            "name": "max_priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 5,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum priority level (1=Critical, 5=Very Low)",
              "title": "Max Priority"
            }
          },
          {
            "description": "Specific priority levels to include",
            "in": "query",
            "name": "priority_levels",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific priority levels to include",
              "title": "Priority Levels"
            }
          },
          {
            "description": "Filter by classification method (bert, rule_based, manual)",
            "in": "query",
            "name": "classification_method",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by classification method (bert, rule_based, manual)",
              "title": "Classification Method"
            }
          },
          {
            "description": "Include priority distribution statistics",
            "in": "query",
            "name": "include_priority_stats",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include priority distribution statistics",
              "title": "Include Priority Stats",
              "type": "boolean"
            }
          },
          {
            "description": "Sort by: start_time, priority, confidence",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "start_time",
              "description": "Sort by: start_time, priority, confidence",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "description": "Sort order: asc, desc",
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "default": "asc",
              "description": "Sort order: asc, desc",
              "title": "Sort Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response List My Events Api V1 Events  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Create a new event for the specified user with optional BERT priority classification\nEnhanced with real-time WebSocket broadcasting\n\nArgs:\n    cognito_sub: User's Cognito sub ID\n    event_data: Event creation data\n    auto_classify_priority: Whether to automatically classify priority using BERT\n    db: Database session\n    \nReturns:\n    Created event with priority classification and real-time notifications",
        "operationId": "create_event_api_v1_events__post",
        "parameters": [
          {
            "in": "query",
            "name": "cognito_sub",
            "required": true,
            "schema": {
              "title": "Cognito Sub",
              "type": "string"
            }
          },
          {
            "description": "Automatically classify event priority using BERT",
            "in": "query",
            "name": "auto_classify_priority",
            "required": false,
            "schema": {
              "default": true,
              "description": "Automatically classify event priority using BERT",
              "title": "Auto Classify Priority",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/users/": {
      "get": {
        "description": "List all users (admin/development only)",
        "operationId": "list_users_api_v1_users__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Api V1 Users  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "description": "Create a new user (simplified - no auth required for now).\n\nRobust duplicate handling & normalization to avoid 500s from constraint violations.",
        "operationId": "create_user_api_v1_users__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}